qqidea/requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.4
django-debug-toolbar==3.8.1
django-rest-framework==0.1.0
django-simpleui==2022.12.18
djangorestframework==3.14.0
executing==1.2.0
idna==3.4
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
pytz==2022.7
PyYAML==6.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
traitlets==5.8.0
uritemplate==4.1.1
urllib3==1.26.13
wcwidth==0.2.5
